@@ -1,7 +1,7 @@
 package accountingtest
 
 import (
-	"math/rand"
+	"math/rand/v2"
 
 	"github.com/nspcc-dev/neofs-sdk-go/accounting"
 )
@@ -9,7 +9,7 @@ import (
 // Decimal returns random accounting.Decimal.
 func Decimal() accounting.Decimal {
 	var d accounting.Decimal
-	d.SetValue(rand.Int63())
+	d.SetValue(rand.Int64())
 	d.SetPrecision(rand.Uint32())
 
 	return d
