@@ -1,7 +1,7 @@
 package containertest
 
 import (
-	"math/rand"
+	"math/rand/v2"
 
 	"github.com/nspcc-dev/neofs-sdk-go/container"
 	"github.com/nspcc-dev/neofs-sdk-go/container/acl"
